+++ collector/connect_godror.go
@@ -20,14 +20,11 @@ func connect(logger *slog.Logger, dbname string, dbconfig DatabaseConfig) (*sql.
 	logger.Debug("Launching connection to "+maskDsn(dbconfig.URL), "database", dbname)
 
 	var P godror.ConnectionParams
-	password, err := dbconfig.GetPassword()
+	credentials, err := dbconfig.ResolveCredentials()
 	if err != nil {
 		return nil, err
 	}
-	username, err := dbconfig.GetUsername()
-	if err != nil {
-		return nil, err
-	}
+	username, password := credentials.Username, credentials.Password
 	// If password is not specified, externalAuth will be true, and we'll ignore user input
 	dbconfig.ExternalAuth = password == ""
 	logger.Debug(fmt.Sprintf("external authentication set to %t", dbconfig.ExternalAuth), "database", dbname)